Minor refactoring related to unneeded classes like Enemies, Characters and Entities. They are now simply part of the Game class. Moved all Enums to the Enum folder. Simplified the levels to make testing easier.
