+++ config/levels.json
@@ -1,6 +1,55 @@
 {
     "data": [
         {
+            "name": "Level 0",
+            "background": "Space3",
+            "characters": [
+                {
+                    "id": "b58db7a6-f8fc-4f31-9162-86fae6282303",
+                    "sprite": "player",
+                    "animationKey": "Player/Default/skeleton-MovingNIdle",
+                    "animationSpeed": 0.14,
+                    "isPlayer": true,
+                    "position": {
+                        "x": 100,
+                        "y": 525
+                    },
+                    "life": 100,
+                    "shield": 300
+                },
+                {
+                    "id": "6f577ac4-4119-41fe-b620-ca26f8a5ebbf",
+                    "sprite": "enemy01",
+                    "animationKey": "Enemy01/Idle/skeleton-Idle",
+                    "animationSpeed": 0.14,
+                    "isPlayer": false,
+                    "position": {
+                        "x": 1300,
+                        "y": 25
+                    },
+                    "life": 100,
+                    "shield": 300
+                },
+                {
+                    "id": "1927c78c-ac9c-4fed-a29b-d0cdb9142054",
+                    "sprite": "enemy02",
+                    "animationKey": "Enemy02/Idle/skeleton-Idle",
+                    "animationSpeed": 0.14,
+                    "isPlayer": false,
+                    "position": {
+                        "x": 1300,
+                        "y": 200
+                    },
+                    "life": 100,
+                    "shield": 300
+                }
+            ],
+            "entities": [
+                {}
+            ],
+            "backgroundMusic": "sfx/music/music_8bit_jammer"
+        },
+        {
             "name": "Level 1",
             "background": "Space1",
             "characters": [
@@ -75,110 +124,6 @@
                     "position": {
                         "x": 1300,
                         "y": 200
-                    },
-                    "life": 100,
-                    "shield": 300
-                },
-                {
-                    "id": "f544cf59-6626-462b-bdc8-2c636c20c21c",
-                    "sprite": "enemy03",
-                    "animationKey": "Enemy03/Idle/skeleton-Idle",
-                    "animationSpeed": 0.14,
-                    "isPlayer": false,
-                    "position": {
-                        "x": 1300,
-                        "y": 400
-                    },
-                    "life": 100,
-                    "shield": 300
-                },
-                {
-                    "id": "502727ec-c140-44f1-afdc-f414b5a55ed6",
-                    "sprite": "enemy04",
-                    "animationKey": "Enemy04/Idle/skeleton-Idle",
-                    "animationSpeed": 0.14,
-                    "isPlayer": false,
-                    "position": {
-                        "x": 1300,
-                        "y": 600
-                    },
-                    "life": 100,
-                    "shield": 300
-                },
-                {
-                    "id": "f2c2f8c4-5ba1-4e25-9438-5db1fd019385",
-                    "sprite": "enemy05",
-                    "animationKey": "Enemy05/Idle/skeleton-Idle",
-                    "animationSpeed": 0.14,
-                    "isPlayer": false,
-                    "position": {
-                        "x": 1300,
-                        "y": 800
-                    },
-                    "life": 100,
-                    "shield": 300
-                },
-                {
-                    "id": "864b497c-8443-4264-818c-3f72b92a3ea2",
-                    "sprite": "enemy06",
-                    "animationKey": "Enemy06/Idle/skeleton-Idle",
-                    "animationSpeed": 0.14,
-                    "isPlayer": false,
-                    "position": {
-                        "x": 1600,
-                        "y": 25
-                    },
-                    "life": 100,
-                    "shield": 300
-                },
-                {
-                    "id": "2be201d2-b110-4142-9610-66738db50833",
-                    "sprite": "enemy07",
-                    "animationKey": "Enemy07/Idle/skeleton-Idle",
-                    "animationSpeed": 0.14,
-                    "isPlayer": false,
-                    "position": {
-                        "x": 1600,
-                        "y": 200
-                    },
-                    "life": 100,
-                    "shield": 300
-                },
-                {
-                    "id": "27277559-0eae-4327-9042-ffb31275b3a2",
-                    "sprite": "enemy08",
-                    "animationKey": "Enemy08/Idle/skeleton-Idle",
-                    "animationSpeed": 0.14,
-                    "isPlayer": false,
-                    "position": {
-                        "x": 1600,
-                        "y": 400
-                    },
-                    "life": 100,
-                    "shield": 300
-                },
-                {
-                    "id": "707cc366-5b28-4aad-a64d-2f26b7c5d7c4",
-                    "sprite": "enemy09",
-                    "animationKey": "Enemy09/Idle/skeleton-Idle",
-                    "animationSpeed": 0.14,
-                    "isPlayer": false,
-                    "position": {
-                        "x": 1600,
-                        "y": 600
-                    },
-                    "life": 100,
-                    "shield": 300
-                },
-                {
-                    "id": "78e15dcb-54e3-4c81-acf3-610b9eee93d4",
-                    "sprite": "enemy10",
-                    "animationKey": "Enemy10/Idle/skeleton-Idle",
-                    "animationSpeed": 0.14,
-                    "isPlayer": false,
-                    "position": {
-                        "x": 1600,
-                        "y": 800
                     },
                     "life": 100,
                     "shield": 300
