+++ ts/Enemies.ts
@@ -4,10 +4,7 @@ import { ActiveActionSprite } from "./ActiveActionSprite";
 import { calculateMovement } from "./Functions";
 import { Game } from "./Game";
 import { Character, CharacterAction } from "./Models/Character";
-
-export class Enemies {
-    data: Map<string, Enemy> = new Map<string, Enemy>();
-}
+import { MoveBox } from "./MoveBox";
 
 export class Enemy {
 
@@ -389,36 +386,5 @@ export class Enemy {
                 }
             }
         }
-    }
-}
-
-export class MoveBox {
-    /**  */
-    constructor(minX: number, maxX: number, minY: number, maxY: number) {
-        this._minX = minX;
-        this._maxX = maxX;
-        this._minY = minY;
-        this._maxY = maxY;
-    }
-
-    private _minX: number;
-    private _maxX: number;
-    private _minY: number;
-    private _maxY: number;
-
-    get minX(): number {
-        return this._minX;
-    }
-
-    get maxX(): number {
-        return this._maxX;
-    }
-
-    get minY(): number {
-        return this._minY;
-    }
-
-    get maxY(): number {
-        return this._maxY;
     }
 }
