+++ ts/Enums/Actions.ts
@@ -1,0 +1,4 @@
+export enum Actions {
+  Move,
+  Fire
+}
