+++ ts/Enums/GameState.ts
@@ -1,0 +1,7 @@
+export enum GameState {
+  Loading,
+  Menu,
+  Paused,
+  Stopped,
+  Running
+}
