+++ ts/Game.ts
@@ -1,14 +1,17 @@
 import { Point } from "pixi.js";
 import { Collision2D } from "./Collision2D";
-import { Enemies, Enemy, MoveBox } from "./Enemies";
+import { Enemy } from "./Enemies";
 import { loadAnimationSprites, loadBackgrounds, loadCharacters, loadEntities, loadLevels, loadSounds } from "./Functions";
 import { AnimationSprites } from "./Models/AnimatedSprite";
 import { Backgrounds } from "./Models/Background";
-import { Character, Characters } from "./Models/Character";
+import { Character } from "./Models/Character";
 import { DrawText } from "./Models/DrawText";
-import { Entities, EntitySound } from "./Models/Entities";
+import { EntitySound, Entity } from "./Models/Entities";
 import { LevelData, Levels } from "./Models/Level";
+import { MoveBox } from "./MoveBox";
 import { Player } from "./Player";
+import { Actions } from "./Enums/Actions";
+import { GameState } from "./Enums/GameState";
 
 export class Game {
 
@@ -25,10 +28,10 @@ export class Game {
   backgrounds: Backgrounds;
   levels: Levels;
   animationSprites: AnimationSprites;
-  entities: Entities;
-  characters: Characters;
+  entities: Map<string, Entity>;
+  characters: Map<string, Character>;
   player: Player;
-  enemies: Enemies;
+  enemies: Map<string, Enemy>;
 
   /** Sounds */
   backgroundMusic: EntitySound;
@@ -182,11 +185,11 @@ export class Game {
     // If a level is already loaded, remove all items from the stage
     if (this.level) {
 
-    // Remove the player from the stage
+      // Remove the player from the stage
       this.player.removeStage();
 
       // Remove all enemies from the stage
-      this.enemies.data.forEach(enemy => {
+      this.enemies.forEach(enemy => {
         enemy.removeStage()
       });
 
@@ -216,8 +219,8 @@ export class Game {
     this.level = this.levels.data[this.levelIndex];
 
     // Reset all enemies
-    this.enemies = new Enemies();
-    this.enemies.data.clear();
+    this.enemies = new Map<string, Enemy>();
+    this.enemies.clear();
 
     // Add all characters to the stage
     for (let i = 0; i < this.level.characters.length; i++) {
@@ -236,7 +239,7 @@ export class Game {
         // Initialize the enemy
         enemy.init();
 
-        this.enemies.data.set(character.id, enemy);
+        this.enemies.set(character.id, enemy);
       }
     }
 
@@ -266,7 +269,7 @@ export class Game {
     }
 
     // Load sounds
-    this.backgroundMusic = this.entities.data.get(this.level.backgroundMusic).sound;
+    this.backgroundMusic = this.entities.get(this.level.backgroundMusic).sound;
     PIXI.sound.play(this.backgroundMusic.id, {
       volume: this.backgroundMusic.volume,
       loop: true
@@ -328,7 +331,7 @@ export class Game {
       }
 
       // Update all enemy
-      this.enemies.data.forEach(enemy => {
+      this.enemies.forEach(enemy => {
         enemy.update(this);
       })
 
@@ -337,7 +340,7 @@ export class Game {
         if (action.triggerEvents) {
 
           // Check collision for all enemies
-          this.enemies.data.forEach(enemy => {
+          this.enemies.forEach(enemy => {
             if (Collision2D.boxedCollision(
               action.sprite.position, enemy.position,
               action.sprite.getLocalBounds(), enemy.character.animation.getLocalBounds()
@@ -376,10 +379,10 @@ export class Game {
                   enemy.removeStage();
 
                   // Remove the enemies from the list
-                  g.enemies.data.delete(enemy.id);
+                  g.enemies.delete(enemy.id);
 
                   // If all enemies are gone, load the next level
-                  if (g.enemies.data.size == 0) {
+                  if (g.enemies.size == 0) {
                     g.levelIndex++;
                     g.loadLevel();
                   }
@@ -430,7 +433,7 @@ export class Game {
       }
 
       // Update all enemies
-      this.enemies.data.forEach(enemy => {
+      this.enemies.forEach(enemy => {
         enemy.moveBox = new MoveBox(this.app.screen.width / 2, this.app.screen.width, 0, this.app.screen.height);
       })
     }
@@ -486,15 +489,3 @@ window.onkeyup = function (e: any) {
   kp[code] = false;
 };
 
-export enum Actions {
-  Move,
-  Fire
-}
-
-export enum GameState {
-  Loading,
-  Menu,
-  Paused,
-  Stopped,
-  Running
-}
